data-ingestion-service/requirements.txt
kafka-python==2.0.2
schedule==1.1.0